rename to extensions https://github.com/MLDSAI/OpenAdapt/pull/318#discussion_r1246073970
